Refactor transport config
